New error handling and launch
Updated version.  Improved error handling.  If a connection error is detected then the application pauses for 30 seconds - clock freezes, then attempts to continue.  It is still possible  for network errors to crash the application.  To allow you to check for this it writes a PID file with the PID number of the running application.  There is an additional program (monclock.py) which can be used to launch weatherclock.py.  To do this run the application in place of weatherclock.py and every minute it should check to see if it is running and restart weatherclock.py if not.